services/webhook-dispatcher/requirements.txt
aiohttp==3.9.1
aio-pika==9.3.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0

